feat(next migration): add compiling CSV on the server
